+++ tinyfaces/evaluate.py
@@ -1,47 +1,13 @@
-import argparse
 import os
 import os.path as osp
 
 import numpy as np
 import torch
 from torchvision import transforms
-from tqdm import tqdm
 
-from tinyfaces.datasets import get_dataloader
 from tinyfaces.models.model import DetectionModel
 from tinyfaces.models.utils import get_bboxes
 from tinyfaces.utils.nms import nms
-
-
-def arguments():
-    parser = argparse.ArgumentParser("Model Evaluator")
-    parser.add_argument("dataset")
-    parser.add_argument("--split", default="val")
-    parser.add_argument("--dataset-root")
-    parser.add_argument("--checkpoint",
-                        help="The path to the model checkpoint",
-                        default="")
-    parser.add_argument("--prob_thresh", type=float, default=0.03)
-    parser.add_argument("--nms_thresh", type=float, default=0.3)
-    parser.add_argument("--workers", default=8, type=int)
-    parser.add_argument("--batch_size", default=1, type=int)
-    parser.add_argument("--results_dir", default=None)
-    parser.add_argument("--debug", action="store_true")
-
-    return parser.parse_args()
-
-
-def dataloader(args):
-    normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
-                                     std=[0.229, 0.224, 0.225])
-    val_transforms = transforms.Compose([transforms.ToTensor(), normalize])
-
-    val_loader, templates = get_dataloader(args.dataset,
-                                           args,
-                                           train=False,
-                                           split=args.split,
-                                           img_transforms=val_transforms)
-    return val_loader, templates
 
 
 def get_model(checkpoint=None, num_templates=25):
@@ -141,57 +107,3 @@ def write_results(dets, img_path, split, results_dir=None):
             d = "{0} {1} {2} {3} {4}\n".format(int(left), int(top), int(width),
                                                int(height), score)
             f.write(d)
-
-
-def run(model,
-        val_loader,
-        templates,
-        prob_thresh,
-        nms_thresh,
-        device,
-        split,
-        results_dir=None,
-        debug=False):
-    for _, (img, filename) in tqdm(enumerate(val_loader),
-                                   total=len(val_loader)):
-        dets = get_detections(model,
-                              img,
-                              templates,
-                              val_loader.dataset.rf,
-                              val_loader.dataset.transforms,
-                              prob_thresh,
-                              nms_thresh,
-                              device=device)
-
-        write_results(dets, filename[0], split, results_dir)
-    return dets
-
-
-def main():
-    args = arguments()
-
-    if torch.cuda.is_available():
-        device = torch.device('cuda:0')
-    else:
-        device = torch.device('cpu')
-
-    val_loader, templates = dataloader(args)
-    num_templates = templates.shape[0]
-
-    model = get_model(args.checkpoint, num_templates=num_templates)
-
-    with torch.no_grad():
-        # run model on val/test set and generate results files
-        run(model,
-            val_loader,
-            templates,
-            args.prob_thresh,
-            args.nms_thresh,
-            device,
-            args.split,
-            results_dir=args.results_dir,
-            debug=args.debug)
-
-
-if __name__ == "__main__":
-    main()
